chore: add comprehensive checkers

+++ modules/directlfq.nf
@@ -12,13 +12,20 @@ process DIRECTLFQ_FORMAT {
     //
 
     script:
-    """
-    mkdir paths; mv $scan_prot_mappings paths
-    Rscript $params.bin/R/directlfq_format2.r \
-        -p paths \
-        -m $msms_mapping \
-        -o directlfq.aq_reformat.tsv
-    """
+    output = "${outdir}/directlfq.aq_reformat.tsv"
+    if (file(output).exists()) {
+        """
+        cp ${output} .
+        """
+    } else {
+        """
+        mkdir paths; mv $scan_prot_mappings paths
+        Rscript $params.bin/R/directlfq_format2.r \
+            -p paths \
+            -m $msms_mapping \
+            -o directlfq.aq_reformat.tsv
+        """
+    }
     //
 }
